Boost on Outdated CMake
Fix the tests on CMake versions that are not aware of a newer
boost release yet.

This is a work-around for releases of Boost until CMake support
becomes mature in Boost. Rely on the PUBLIC link & include
of core for tests if the target for the unit system is not found.