Container Particle Species: Fix Overload
The overload of the `Container::operator[]` can not be done by
re-defining the symbol again a second time. This surprisingly works
with the Linux/OSX linkers but is hacky.

Instead, we can define a trait to pass additional modifications after
creation/emplacement into a container.

As a nice side-effect, this also removes code-duplication.

Fixes a linker issue on MSVC as well.