Docs: MPI Version
We currently even test on OpenMPI 1.6.5 which only
provides MPI 2.1

Of course we can require something more recent anytime we need it.
This just means we have to compile it or use the mpich2 libs on
trusty (or better: both, in order to continue testing the oldest
claimed supported version as well).