Python: Use Lists for Offset & Extents
Offset & Extent classes as std::vector< > types are notoriously
hard to expose properly and make transparent since they are all
just named aliases of the same type that is also pre-defined in
pybind11.

Anyway, it is actually way more natural to use standard python
lists for them. There are even casts already defined from python
lists to std::initializer_lists and vectors, so we can keep the
types in the interface.