Travis: Skip clang-tidy & IWYU
Although both tools are very promising, we are currently not fixing
their reported issues to 100%.

That would be no problem to fix them gradually, but since they do not
succeed yes, travis is not caching our (dependency) build artifacts.
Due to that, both take up 20min and block development time.

Proposal how to enable them: in case there are no false positives or
we can add config files, e.g. for `clang-format` we also need to add
a config file, a PR should fix issues caused by them and enable them
in the process (or later).