requirements.txt: Upper bounds
New major releases of these dependencies would very likely cause
problems and were not tested on release of this software. If they
work, we can bump the bound (in packages recipes aka package builds).

For the bare software install, we don't define those bounds to
allow bumping them in packaging alone without the need for a new
mainline release.