CMake: Optional ADIOS1 Wrapper
We need to build and install the optional libs to wrap
ADIOS1 (hide doubly defined MPI mock symbols in serial lib)
only if we actually enable ADIOS1 support.

Otherwise, we can also throw the missing support warning
early on, without the need for implementation files and
auxiliary libraries that wrap a dummy ADIOS1.