JSON: Always Enabled
Since we ship the dependencies for the private JSON backend with
our source code and the dependency is of very hight quality and
very portable, we can just compile it by default.

This is very similar to previous behavior, in the sense that it
was `-DopenPMD_USE_JSON=AUTO` (internal) activated unless explicitly
deactivated during compile.

Enabling JSON by default and making it a required dependency enables
us to use it for further internal tasks, such as runtime option
parsing of user parameters for all backends.