CMake: Enable IPO/LTO
Interprocedural optimization (IPO) also known as link-time optimization (LTO)
is a great feature to reduce binary size and to improve code performance, e.g.
through extended inlining.

This feature is specifically useful for modern C++ code, which is also the
reason pybind11 enables it automatically for our python bindings as soon as
it is detected.

We now expose this feature for the rest of the library build as well and
auto-enable it when supported by the compiler.

Binary size before:
```
$ du -hs lib/*
1.0M    lib/libCatchMain.so
1.0M    lib/libCatchRunner.so
1.3M    lib/libopenPMD.ADIOS1.Parallel.so
1.1M    lib/libopenPMD.ADIOS1.Serial.so
2.2M    lib/libopenPMD.so
968K    lib/python3.6
```
and with IPO/LTO enabled (new default):
```
$ du -hs lib/*
876K    lib/libCatchMain.so
876K    lib/libCatchRunner.so
1.2M    lib/libopenPMD.ADIOS1.Parallel.so
1.1M    lib/libopenPMD.ADIOS1.Serial.so
1.5M    lib/libopenPMD.so
968K    lib/python3.6
```