setup.py / pip: Static Build by default
The default build for `setup.py`/`pip`/`wheel` builds is now changed
to static. This solves install problems on Windows, where we cannot
set RPath hints.

Unix builds are fine both shared and static and are the only ones
that need RPaths, to find the (optional) ADIOS1 MPI/non-MPI wrapper
libraries.

The Python `.so`/`.pyd` are still shared libraries (as they have to
be).