ADIOS1: Keep Default for Now
Since we are planning to add more refactoring in the internal
(variables vs. attributes) representation of ADIOS2, let's not
make this default over ADIOS1 yet to avoid writing files that
cannot be read again soon.

Delay this just for a little for this upcoming release, so we
can draft a new version without changing defaults yet. (I want
to release a new version very soon to fix some installation
issues and other updates that we accumulated.)