NVCC: No Visibility Attribute
This project is provides host-side APIs [1]. Some downstream
projects are unable to split their translation units properly in
host+device and host-only, and just pass all code through `nvcc`.
Doing so is not appreciated (by me) but tolerated (by me).

Unfortunately, `nvcc` is a bit behind in supporting visibility
attributes (Nvidia bug report: 2767443 as of v10.2). Assuming that
openPMD-api itself is built with a proper host-compiler instead of
`nvcc`, we can live with defining our `EXPORT` macro to nothing
during downstream consumption.

This causes a compile error when build with GCC <6 and a warning
with newer GCC in `include/openPMD/IO/IOTask.hpp` for
`enum class Operation`.

[1] Note: We might support passing device-side pointers in the
future, though.