Docs: All Contributors
Collecting all (smaller but valuable) code and non-code contributions
to the library.

I am not sure if I am able to keep up with the details or if a simple
list of names is easier, but at least we can try. Thank you all for
contributing, improving, testing and reporting in various ways!