Defaults for Date & Software Attributes
Adds defaults for the date and software attributes, which are
recommended openPMD attributes (missing is a warning in our
validator).

Deprecate old API that allows to set the software version independent
of the software name, which can lead to inconsistencies, e.g. if
the default version is kept and only the software name is changed.