Add TracingJSON class
This will allow warning users of unused entries in the JSON config.

Invert TracingJSON shadow logic

Instead of removing all accessed entries one after another, add them to
the "shadow" value one after another. This is a bit more tricky since
the result has to be inverted again in the end, but it is less prone to
creating dangling pointer situations.