Test Case: FBPIC Empty HDF5
FBPIC writes solely to HDF5 and chooses to write a variable with
empty extent for zero particles. Let's test this case.

Note: we cannot create this with the API on the fly, since we use
the more portable "constant record component for empty" approach
that works with all file formats.