Fix & Replace Python Data_Order
The `Data_Order` enum in python and thus `Mesh::data_order` are
broken due to an upstream bug in pybind11 with non-int enums.

Replace the setter and read-only property with a unified property
and remove the enum for the sake of simplicity (just compare to a
char in popular Python "enums are strings, better spell it right"-
manner).