CI: Brew Returns Non-Zero If Already Installed
Homebrew has the unusual behavior to return a non-zero error code in
`brew install` if the package is already installed. Since the base
image in CI is somewhat fluid and also the status that `brew update`
leaves is very dynamic, we just change this ourselves now to be a
non-error.