Add distinction of three flush levels
UserFlush: Exposed flush point, buffers are read from / written to

FlushEverything (todo: rename to InternalFlush): Flush everything
except for things that are only allowed to be flushed at a flush point.
Attributes must be flushed.

SkeletonOnly: Guarantees to setup the openPMD hierarchy in the backends.
Datasets are not created yet.