Fix: macOS ADIOS1 w/o XCode
Some compiler toolchains on macOS do not use the XCode system
compiler. For instance, conda-forge uses vanilla clang to build
packages.

In that case, the symbol hiding instructions we used on macOS
were not effective, leading to errors like #1195.

The current patch checks if the toolchain understands GCC/LLVM-ish
linker flags on macOS and if it does, it adds them.