Frontend implementation of appending
Also use switch statement in many frontend places for the Access enum.
With 4 variants, if statements become increasingly unreadable and it's
good to have compiler warnings if a case is forgotten. Try to avoid
default branches.

Also refactor `autoDetectPadding()` into a separate function since it is
now needed in two places.